+++ src/Twig/Components/FormationParcoursComponent.php
@@ -8,7 +8,7 @@ use Symfony\UX\LiveComponent\Attribute\AsLiveComponent;
 use Symfony\UX\LiveComponent\Attribute\LiveProp;
 use Symfony\UX\LiveComponent\DefaultActionTrait;
 use Symfony\UX\TwigComponent\Attribute\PostMount;
-
+//@deprecated
 #[AsLiveComponent('formation_parcours')]
 final class FormationParcoursComponent
 {