+++ src/Twig/Components/MentionManageComponent.php
@@ -3,11 +3,14 @@
 namespace App\Twig\Components;
 
 use App\Classes\GetDpeParcours;
+use App\Classes\MentionProcess;
 use App\Classes\ValidationProcess;
 use App\Entity\FicheMatiere;
 use App\Entity\Formation;
 use App\Entity\Parcours;
 use App\Entity\TypeDiplome;
+use App\Enums\EtatDemandeChangeRfEnum;
+use App\Enums\EtatProcessMentionEnum;
 use App\Enums\TypeModificationDpeEnum;
 use App\Repository\HistoriqueFormationRepository;
 use App\Repository\HistoriqueParcoursRepository;
@@ -70,14 +73,8 @@ final class MentionManageComponent extends AbstractController
     #[LiveProp(writable: true)]
     public ?Formation $formation = null;
 
-    #[LiveProp]
-    public ?Parcours $parcours = null;
-
     public array $process;
     public array $historiques = [];
-
-    #[LiveProp]
-    public ?FicheMatiere $ficheMatiere = null;
 
     #[LiveProp]
     public ?TypeDiplome $typeDiplome;
@@ -95,7 +92,7 @@ final class MentionManageComponent extends AbstractController
     public function __construct(
         private readonly HistoriqueFormationRepository $historiqueFormationRepository,
         private readonly HistoriqueParcoursRepository  $historiqueParcoursRepository,
-        private readonly ValidationProcess             $validationProcess,
+        private readonly MentionProcess                $validationProcess,
         #[Target('dpeParcours')]
         private readonly WorkflowInterface             $dpeParcoursWorkflow,
     ) {
@@ -115,13 +112,6 @@ final class MentionManageComponent extends AbstractController
 
     public function redirige(): Response
     {
-        // si niveau parcours, formation ou composante, une fois validé => On redirige vers le show.
-        if ($this->type === 'parcours') {
-            return $this->redirectToRoute('app_parcours_show', [
-                'id' => $this->parcours->getId()
-            ]); //uniquement si RF ou DPE
-        }
-
         return $this->redirectToRoute('app_formation_show', [
             'slug' => $this->formation->getSlug()
         ]); //uniquement si RF ou DPE
@@ -155,54 +145,25 @@ final class MentionManageComponent extends AbstractController
 
     private function getHistorique(): void
     {
-        if ($this->type === 'formation') {
-            $historiques = $this->historiqueFormationRepository->findBy(['formation' => $this->formation], ['created' => 'ASC']);
-            foreach ($historiques as $historique) {
-                $this->historiques[$historique->getEtape()] = $historique;
-            }
-        } elseif ($this->type === 'parcours') {
-            $historiques = $this->historiqueParcoursRepository->findBy(['parcours' => $this->parcours], ['created' => 'ASC']);
-
-            foreach ($historiques as $historique) {
-                if (self::TAB_PROCESS[$historique->getEtape()] < self::TAB_PROCESS[self::TAB[$this->etape]]) {
-                    $this->historiques[$historique->getEtape()] = $historique;
-                }
-            }
-            $historiques = $this->historiqueFormationRepository->findBy(['formation' => $this->parcours->getFormation()], ['created' => 'ASC']);
-            foreach ($historiques as $historique) {
-                if (self::TAB_PROCESS[$historique->getEtape()] < self::TAB_PROCESS[self::TAB[$this->etape]]) {
-                    $this->historiques[$historique->getEtape()] = $historique;
-                }
-            }
-
+        $historiques = $this->historiqueFormationRepository->findBy(['formation' => $this->formation], ['created' => 'ASC']);
+        foreach ($historiques as $historique) {
+            $this->historiques[$historique->getEtape()] = $historique;
         }
-
     }
 
     #[PostMount]
     public function postMount(): void
     {
-        if ($this->type === 'formation') {
-            $this->typeDiplome = $this->formation->getTypeDiplome();
-            $this->place = $this->getPlace($this->type);
+        $this->typeDiplome = $this->formation->getTypeDiplome();
+        $this->place = $this->getPlace($this->type);
 
-            if ($this->formation !== null && $this->formation->isHasParcours() === false && $this->formation->getParcours()->count() === 1) {
+        if ($this->formation !== null && $this->formation->isHasParcours() === false && $this->formation->getParcours()->count() === 1) {
 
-                $parcours = $this->formation->getParcours()->first();
-                $this->hasDemande =
-                    GetDpeParcours::getFromParcours($parcours)?->getEtatReconduction() === TypeModificationDpeEnum::MODIFICATION_TEXTE || GetDpeParcours::getFromParcours($parcours)?->getEtatReconduction() === TypeModificationDpeEnum::MODIFICATION_MCCC_TEXTE || GetDpeParcours::getFromParcours($parcours)?->getEtatReconduction() === TypeModificationDpeEnum::MODIFICATION_MCCC;
-            }
-        } elseif ($this->type === 'parcours' && $this->parcours !== null) {
-            $this->typeDiplome = $this->parcours?->getFormation()?->getTypeDiplome();
-            $this->formation = $this->parcours?->getFormation();
-            $this->place = $this->getPlace($this->type);
+            $parcours = $this->formation->getParcours()->first();
             $this->hasDemande =
-                GetDpeParcours::getFromParcours($this->parcours)?->getEtatReconduction() === TypeModificationDpeEnum::MODIFICATION_TEXTE || GetDpeParcours::getFromParcours($this->parcours)?->getEtatReconduction() === TypeModificationDpeEnum::MODIFICATION_MCCC_TEXTE || GetDpeParcours::getFromParcours($this->parcours)?->getEtatReconduction() === TypeModificationDpeEnum::MODIFICATION_MCCC;
-        } elseif ($this->type === 'ficheMatiere') {
-            $this->typeDiplome = $this->ficheMatiere->getParcours()?->getFormation()?->getTypeDiplome();
-            $this->parcours = $this->ficheMatiere->getParcours();
-            $this->formation = $this->parcours->getFormation();
+                GetDpeParcours::getFromParcours($parcours)?->getEtatReconduction() === TypeModificationDpeEnum::MODIFICATION_TEXTE || GetDpeParcours::getFromParcours($parcours)?->getEtatReconduction() === TypeModificationDpeEnum::MODIFICATION_MCCC_TEXTE || GetDpeParcours::getFromParcours($parcours)?->getEtatReconduction() === TypeModificationDpeEnum::MODIFICATION_MCCC;
         }
+
         $this->etape = self::TAB[$this->place] ?? $this->type;
         $this->getHistorique();
 
@@ -227,11 +188,197 @@ final class MentionManageComponent extends AbstractController
 
     private function getPlace(string $type): string
     {
-        $objet = $type === 'formation' || $this->parcours === null ? GetDpeParcours::getFromFormation($this->formation) : GetDpeParcours::getFromParcours($this->parcours);
+        $objet = GetDpeParcours::getFromFormation($this->formation);
         if (null === $objet) {
             return 'initialisation_dpe';
         }
 
         return array_keys($this->getWorkflow()->getMarking($objet)->getPlaces())[0];
+    }
+
+    public function stateProcess(string $etat): EtatProcessMentionEnum
+    {
+        /*
+         * Etat des fiches
+            bleu - au moins une fiche en cours de rédaction et non validée
+            orange - réserve sur au moins une fiche ou non conforme
+            vert - Toutes les fiches sont validées, pas de soucis
+         * Etat des parcours
+            bleu - au moins un DPE parcours en cours de rédaction et non validée CFVU (en cours de process)
+            orange - réserve sur au moins une fiche DPE parcours ou sans PV associé ou non conforme
+            vert - Toutes les fiches DPE parcours sont validées CFVU
+         * Etat formation
+            bleu - au moins un DPE en cours de rédaction et non validée SES (en cours de process) - vérifier qu'il n'y a pas de modifications de textes en cours
+            orange - réserves sur fiche DPE ou non conforme
+            vert - fiche DPE validées SES
+         * Change RF
+            bleu - le RF en cours de modification et non validée CFVU (en cours de process)
+            orange - réserve sur le RF ou sans PV associé
+            vert - RF est validé CFVU avec PV
+         * Publication
+            bleu - le DPE (formation ou parcours) en cours de modification et non validée à publier (en cours de process)
+            orange - réserve (est-ce utile ?)
+            vert - tous les DPE parcours sont à l'état publication + formation OK
+         */
+        switch ($etat) {
+            case 'fiche_matiere':
+                return $this->getEtatFichesMatieres();
+            case 'parcours':
+                return $this->getEtatParcours();
+            case 'formation':
+                return $this->getEtatFormation();
+            case 'change_rf':
+                return $this->getEtatChangeRf();
+            case 'publication':
+                return $this->getEtatPublication();
+            default:
+                return EtatProcessMentionEnum::WIP;
+        }
+    }
+
+    private function getEtatParcours(): EtatProcessMentionEnum
+    {
+        // * Etat des parcours
+        //            bleu - au moins un DPE parcours en cours de rédaction et non validée CFVU (en cours de process)
+        //            orange - réserve sur au moins une fiche DPE parcours ou sans PV associé ou non conforme
+        //            vert - Toutes les fiches DPE parcours sont validées CFVU
+        $etatsReserves = [
+            'valider_reserve_conseil_cfvu',
+            'reserver_parcours_rf',
+            'reserver_rf',
+            'reserver_dpe_composante',
+            'reserver_conseil',
+            'reserver_central',
+            'valider_reserve_cfvu',
+            'valider_reserve_conseil_cfvu',
+            'valider_reserve_central_cfvu',
+        ];
+        list($states, $nbParcours) = $this->getEtatsParcours();
+
+        foreach ($states as $etat => $nb) {
+            if ($etat === 'autorisation_saisie' && $nb === $nbParcours) {
+                return EtatProcessMentionEnum::NON_FAIT;
+            }
+
+            if (in_array($etat, $etatsReserves)) {
+                return EtatProcessMentionEnum::RESERVE;
+            }
+
+            if ($etat === 'valide_cfvu' && $nb === $nbParcours) {
+                return EtatProcessMentionEnum::COMPLETE;
+            }
+        }
+
+
+        return EtatProcessMentionEnum::WIP;
+    }
+
+    private function getEtatFichesMatieres(): EtatProcessMentionEnum
+    {
+        //* Etat des fiches
+        //            bleu - au moins une fiche en cours de rédaction et non validée
+        //            orange - réserve sur au moins une fiche ou non conforme
+        //            vert - Toutes les fiches sont validées, pas de soucis
+        $nbFiches = 0;
+        $nbFichesPubliees = 0;
+
+
+        foreach ($this->formation->getParcours() as $parcours) {
+            //compter le nombre de fiches matières validées
+            $etatsFiche = $parcours->getEtatsFichesMatieres();
+            $nbFiches += $etatsFiche->nbFiches;
+            $nbFichesPubliees += $etatsFiche->nbFichesPubliees;
+        }
+
+        if ($nbFiches === 0) {
+            return EtatProcessMentionEnum::NON_FAIT;
+        }
+
+        if ($nbFiches === $nbFichesPubliees) {
+            return EtatProcessMentionEnum::COMPLETE;
+        }
+
+        return EtatProcessMentionEnum::WIP;
+    }
+
+    private function getEtatFormation(): EtatProcessMentionEnum
+    {
+        //* Etat formation
+        //            bleu - au moins un DPE en cours de rédaction et non validée SES (en cours de process) - vérifier qu'il n'y a pas de modifications de textes en cours
+        //            orange - réserves sur fiche DPE ou non conforme
+        //            vert - fiche DPE validées SES
+        return EtatProcessMentionEnum::WIP;
+    }
+
+    private function getEtatChangeRf(): EtatProcessMentionEnum
+    {
+        //* Change RF
+        //            bleu - le RF en cours de modification et non validée CFVU (en cours de process)
+        //            orange - réserve sur le RF ou sans PV associé
+        //            vert - RF est validé CFVU avec PV
+        foreach ($this->formation->getChangeRves() as $changeRf) {
+            if ($changeRf->getEtatDemande() === EtatDemandeChangeRfEnum::EN_ATTENTE) {
+                return EtatProcessMentionEnum::WIP;
+            }
+
+            if ($changeRf->getEtatDemande() === EtatDemandeChangeRfEnum::REFUSE || $changeRf->getFichierPv() === null) {
+                //voir si PV passe dans l'historique
+                return EtatProcessMentionEnum::RESERVE;
+            }
+        }
+
+        return EtatProcessMentionEnum::COMPLETE;
+    }
+
+    private function getEtatPublication(): EtatProcessMentionEnum
+    {
+        // * Publication
+        //            bleu - le DPE (formation ou parcours) en cours de modification et non validée à publier (en cours de process)
+        //            orange - réserve (est-ce utile ?)
+        //            vert - tous les DPE parcours sont à l'état publication + formation OK
+        list($states, $nbParcours) = $this->getEtatsParcours();
+
+        foreach ($states as $etat => $nb) {
+            if ($etat === 'autorisation_saisie') {
+                return EtatProcessMentionEnum::NON_FAIT;
+            }
+
+            if ($etat !== 'publie') {
+                return EtatProcessMentionEnum::WIP;
+            }
+
+            if ($etat === 'publie' && $nb === $nbParcours) {
+                //todo: vérifier formation
+                return EtatProcessMentionEnum::COMPLETE;
+            }
+        }
+
+        return EtatProcessMentionEnum::WIP;
+    }
+
+    /**
+     * @return array
+     */
+    private function getEtatsParcours(): array
+    {
+        $states = [];
+        $nbParcours = $this->formation->getParcours()->count();
+        //parcours l'ensemble des parcours de la formation et regarde selon l'état
+        foreach ($this->formation->getParcours() as $parcours) {
+            $objet = GetDpeParcours::getFromParcours($parcours);
+            if (null !== $objet) {
+                $etat = array_keys($this->getWorkflow()->getMarking($objet)->getPlaces())[0];
+
+                if (!array_key_exists($etat, $states)) {
+                    $states[$etat] = 0;
+                }
+
+                $states[$etat]++;
+
+            }
+        }
+
+        ksort($states);
+        return array($states, $nbParcours);
     }
 }