+++ src/TypeDiplome/Export/ButMcccVersion.php
@@ -332,7 +332,7 @@ class ButMcccVersion
         $this->excelWriter->setActiveSheetIndex(0);
         $this->excelWriter->setSelectedCells('A1');
 
-        $this->fileName = Tools::FileName('MCCC - ' . $anneeUniversitaire->getLibelle() . ' - ' . $formation->gettypeDiplome()?->getLibelleCourt() . ' ' . $parcours->getDisplay(), 40);
+        $this->fileName = Tools::FileName('MCCC - ' . $anneeUniversitaire->getLibelle() . ' - ' . $formation->gettypeDiplome()?->getLibelleCourt() . ' ' . $formation->getSigle(). ' ' . $parcours->getSigle(), 40);
     }
 
     public function exportExcelbutMccc(